+++ mods/Smithing/mod/data.lua
@@ -7,11 +7,13 @@ require("prototypes.item.cables")
 
 require("prototypes.anvil")
 
+require("prototypes.recipe.armour_set_recipe")
 require("prototypes.recipe.armour_recipe")
 require("prototypes.recipe.bar_recipe")
 require("prototypes.recipe.intermediate_recipe")
 require("prototypes.recipe.pickaxe_recipe")
 
+require("prototypes.technology.armour_set_technology")
 require("prototypes.technology.armour_technology")
 require("prototypes.technology.processing_technology")
 require("prototypes.technology.pickaxe_technology")